Fix "no such file" error when cleaning thumbnails
The process explorer.exe is using the files and causes an odd symptom.
The first time the files preview and delete normally. The second time
the files will preview, but delete fails with the message
"[Errno 2] No such file or directory."

The new procedure is to kill explorer.exe, clean the files, and then
restart explorer.exe.

https://bugs.launchpad.net/bleachbit/+bug/1571440
reported by ROCKNROLLKID