.desktop and AppStream specifications use reverse-DNS filenames
I don't think there are a lot of applications trying to use the name
bleachbit, but in principle this helps avoid file clobbering.

Also, note that the AppStream specification no longer uses the filename
to distinguish between applications and addons, so all metadata files
now use the .metainfo.xml extension.

<https://specifications.freedesktop.org/desktop-entry-spec/latest/ar01s02.html>

<https://www.freedesktop.org/software/appstream/docs/chap-Quickstart.html#sect-Quickstart-DesktopApps>