migrate where possible from external mock package to stdlib unittest.mock
It was moved to the stdlib in python 3.3, and the PyPI version is a
backport for python versions less than 3.3.

A couple of features from newer python versions that are greater than
the minimum supported version are used. Add if/else guards to utilize
the backport on older versions of python:

- assert_called_once was introduced in python 3.6

- Unstarted patchers were introduced in python 3.8