Fix toolbar and title bar missing
# Symptom
On Microsoft Windows with new GTK 3.24.34, the titlebar and toolbar would sometimes be missing when the application started.

This happened only when the setting was enabled to restore geometry.

# Cause
This was caused by incorrect events reported from GDK, and the application saved the window geometry. This was not visible to the user until later when starting the application.

Closes https://github.com/bleachbit/bleachbit/issues/1753